Split source DB write module
